Added more classifiers and return the average result.
We did some analysis and found out that the score varies from 6-7 to 13-15 when
we use a single RFClassifier. So we thought if we generate a 100 and let them
predict a 100 times each and take the average prediction it should smooth things
out. Which it did. The average error is now between 8 and 9.